Move responsible check to remote instance
