+++ src/domain/notifications/entities/PullRequestNotification.ts
@@ -1,5 +1,4 @@
 import { generateUid } from "d2/uid";
-import { NamedRef } from "../../common/entities/Ref";
 import { MetadataPackage } from "../../metadata/entities/MetadataEntities";
 import { SynchronizationType } from "../../synchronization/entities/SynchronizationType";
 import { BaseNotification } from "./Notification";
@@ -26,10 +25,6 @@ export interface SentPullRequestNotification extends PullRequestNotification {
 export interface ReceivedPullRequestNotification extends PullRequestNotification {
     type: "received-pull-request";
     payload: MetadataPackage;
-    responsibles: {
-        users: NamedRef[];
-        userGroups: NamedRef[];
-    };
 }
 
 export class SentPullRequestNotification implements SentPullRequestNotification {
