+++ src/domain/synchronization/usecases/CreatePullRequestUseCase.ts
@@ -1,4 +1,3 @@
-import _ from "lodash";
 import { cache } from "../../../utils/cache";
 import { NamedRef } from "../../common/entities/Ref";
 import { UseCase } from "../../common/entities/UseCase";
@@ -6,7 +5,6 @@ import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Instance } from "../../instance/entities/Instance";
 import { InstanceRepositoryConstructor } from "../../instance/repositories/InstanceRepository";
 import { MetadataPackage } from "../../metadata/entities/MetadataEntities";
-import { MetadataResponsible } from "../../metadata/entities/MetadataResponsible";
 import { AppNotification, MessageNotification } from "../../notifications/entities/Notification";
 import {
     ReceivedPullRequestNotification,
@@ -41,7 +39,6 @@ export class CreatePullRequestUseCase implements UseCase {
     }: CreatePullRequestParams): Promise<void> {
         const owner = await this.getOwner();
         const { users, userGroups } = notificationUsers;
-        const responsibles = await this.getResponsibles(instance, ids);
 
         const receivedPullRequest = ReceivedPullRequestNotification.create({
             subject,
@@ -53,7 +50,6 @@ export class CreatePullRequestUseCase implements UseCase {
             syncType: type,
             selectedIds: ids,
             payload,
-            responsibles,
         });
 
         const sentPullRequest = SentPullRequestNotification.create({
@@ -100,26 +96,6 @@ export class CreatePullRequestUseCase implements UseCase {
     private async getOwner(): Promise<NamedRef> {
         const { id, name } = await this.instanceRepository(this.localInstance).getUser();
         return { id, name };
-    }
-
-    private async getResponsibles(instance: Instance, ids: string[]) {
-        const responsibles = await this.storageRepository(instance).listObjectsInCollection<
-            MetadataResponsible
-        >(Namespace.RESPONSIBLES);
-
-        const metadataResponsibles = responsibles.filter(({ id }) => ids.includes(id));
-
-        const users = _.uniqBy(
-            metadataResponsibles.flatMap(({ users }) => users),
-            "id"
-        );
-
-        const userGroups = _.uniqBy(
-            metadataResponsibles.flatMap(({ userGroups }) => userGroups),
-            "id"
-        );
-
-        return { users, userGroups };
     }
 
     private async sendMessage(
